Create amino acid base for validators
