+++ variant/validators/amino_acid_delins.py
@@ -9,6 +9,7 @@ from variant.schemas.token_response_schema import Token
 from variant.tokenizers import GeneSymbol
 from variant.tokenizers.caches import AminoAcidCache
 from variant.data_sources import SeqRepoAccess, TranscriptMappings
+from .amino_acid_base import AminoAcidBase
 import logging
 
 
@@ -32,6 +33,7 @@ class AminoAcidDelIns(Validator):
         """
         super().__init__(seq_repo_access, transcript_mappings, gene_symbol)
         self._amino_acid_cache = amino_acid_cache
+        self.amino_acid_base = AminoAcidBase(seq_repo_access, amino_acid_cache)
 
     def get_transcripts(self, gene_tokens, classification, errors)\
             -> Optional[List[str]]:
@@ -81,12 +83,14 @@ class AminoAcidDelIns(Validator):
 
                 # Check ref start/end protein matches expected
                 if not errors:
-                    self.check_ref_aa(t, s.start_aa_del,
-                                      s.start_pos_del, errors)
+                    self.amino_acid_base.check_ref_aa(
+                        t, s.start_aa_del, s.start_pos_del, errors
+                    )
 
                     if not errors and s.end_aa_del:
-                        self.check_ref_aa(t, s.end_aa_del,
-                                          s.end_pos_del, errors)
+                        self.amino_acid_base.check_ref_aa(
+                            t, s.end_aa_del, s.end_pos_del, errors
+                        )
 
                 if not errors and allele not in valid_alleles:
                     results.append(self.get_validation_result(
@@ -104,27 +108,6 @@ class AminoAcidDelIns(Validator):
         # Now add MANE transcripts to result
         self.add_mane_transcript(classification, results, gene_tokens,
                                  mane_transcripts_dict)
-
-    def check_ref_aa(self, t, aa, pos, errors):
-        """Check that reference amino acid matches actual amino acid.
-
-        :param string t: Transcript
-        :param str aa: Expected Amino Acid
-        :param str pos: Expected position
-        :param list errors: List of errors
-        """
-        ref_aa_del = \
-            self.seqrepo_access.sequence_at_position(t, pos)
-        if ref_aa_del and len(ref_aa_del) == 1 \
-                and len(aa) == 3:
-            ref_aa_del = \
-                self._amino_acid_cache.amino_acid_code_conversion[
-                    ref_aa_del]  # noqa: E501
-
-        if ref_aa_del != aa:
-            errors.append(f"Needed to find {aa} at "
-                          f"position {pos} on {t} "
-                          f"but found {ref_aa_del}")
 
     def get_hgvs_expr(self, classification, t, s, is_hgvs) -> str:
         """Return HGVS expression and whether or not it's an Ensembl transcript
